Rename testQuestions.js -> testTestCourseQuestions.js
